+++ geotrek/feedback/forms.py
@@ -1,8 +1,10 @@
 from crispy_forms.layout import Div
 from django.conf import settings
 from django.forms.fields import CharField
-from django.forms.widgets import Textarea
+from django.forms.widgets import HiddenInput, Textarea
+from django.utils.translation import gettext_lazy as _
 
+from geotrek.authent.models import SelectableUser
 from geotrek.common.forms import CommonForm
 
 from .models import Report
@@ -23,6 +25,7 @@ class ReportForm(CommonForm):
             "locked",
             "uid",
             "origin",
+            "assigned_user"
         )
     ]
 
@@ -39,12 +42,14 @@ class ReportForm(CommonForm):
             "locked",
             "uid",
             "origin",
+            "assigned_user"
         ]
         model = Report
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         if settings.SURICATE_MANAGEMENT_ENABLED:
+            self.fields["assigned_user"].queryset = SelectableUser.objects.filter(userprofile__isnull=False)
             self.old_status_id = None
             self.fields["message"] = CharField(required=False)
             self.fields["message"].widget = Textarea()
@@ -54,6 +59,8 @@ class ReportForm(CommonForm):
                 self.old_status_id = self.instance.status.suricate_id
                 self.fields["status"].empty_label = None
                 self.fields["status"].queryset = self.instance.next_status()
+        else:
+            self.fields["assigned_user"].widget = HiddenInput()
 
     def save(self, *args, **kwargs):
         report = super().save(self, *args, **kwargs)